feat(permissions): moved dto and mocks to seperate folders
