+++ src/modules/squeeze/squeeze.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Post, Request } from '@nestjs/common';
 import { ApiOperation, ApiTags } from '@nestjs/swagger';
 import { SqueezeService } from './squeeze.service';
 import { SqueezeRequestDto } from './dto/squeeze.dto';
-import { skipAuth } from 'src/helpers/skipAuth';
+import { skipAuth } from '../../helpers/skipAuth';
 
 @ApiTags('Squeeze')
 @Controller('squeeze')
